api/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
